+++ controllers/manufacturerController.js
@@ -281,6 +281,18 @@ exports.manufacturerDeletePost = asyncHandler(async function (req, res, next) {
   // Redirect to Book List if there is no book to delete
   if (theManufacturer === null) res.redirect("/manufacturer");
 
+  // Check secret key
+  if (req.body.secretKey !== process.env.SECRECT_KEY) {
+    console.log("Secretkey matched");
+    res.render("delete/no_delete_auth", {
+      title: "This is the Category Delete GET page",
+      text: `You can't delete delete '${theManufacturer.name}'`,
+      url: theManufacturer.url,
+      msg: "Sorry the key you provided is incorrect. You can't delete/ update this page. Please refresh your page and try again.",
+    });
+    return;
+  }
+
   // Delete object and redirect to the list of books.
   await Manufacturer.findByIdAndRemove(req.params.id);
   res.redirect("/manufacturer");
